Add files via upload
Intro2CV Final Project Proposal
By Tigran Ramazyan, Evgeny Gurov, Daniil Sobolev

The scientific development of High Energy Physics (HEP) research is closely related to the development of Machine Learning (ML), with the first implementation dated back to 1988. Over the last decade hundreds of papers on usage of ML techniques for accelerating HEP research have been published. Nearly a third of such papers use CV techniques such as CNNs for numerous HEP-ML tasks. That is because any HEP detector is essentially a camera. But instead of making images in the classical notion of the term, it makes images of subnuclear particle collisions. One of the common HEP-ML tasks is classification of high energy electromagnetic showers.

Correctly identifying the nature and properties of particles produced at the Large Hadron Collider is a crucial task for fully exploring subnuclear length scales in search of new physical phenomena. The final project will address the active research problem of object identification in the context of elementary particles interacting with a multi-layer, heterogeneously-segmented electromagnetic calorimeter. A DenseNet-style model should achieve state-of-the-art performance on the e+ - + classification task. The NN approach will be validated by a comparison with other CV methods.

Dataset description can be found by the following link: http://dx.doi.org/10.17632/5fnxs6b557.2 . Image examples can be found on the next page.

We would like to share with you and with the class a “non-traditional” application of CNNs, including the motivation and the methodology, that has been proven to be extremely successful for HEP related tasks.

This ipynb and a list of data represent an algorythm which predicts and evaluates identfied positrons and pions on the pictures from Calorymeter.